requirements_dev.txt
bumpversion==0.5.3
wheel==0.29.0
Django>=1.9.6

django-braces==1.13.0
django-log-request-id==1.3.2
django-cors-headers==3.0.1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.4
djangorestframework-xml==1.4.0
Jinja2==2.10.1
jsonpatch==1.23
jsonpointer==2.0
jsonschema==3.0.1
MarkupSafe==1.1.1
oauthlib==3.0.1
PyYAML==5.1
swagger-parser==1.0.1
swagger-spec-validator==2.4.3
boto3==1.9.146
future==0.17.1
